requirements.txt
fastapi==0.73.0
Flask==2.0.3
numpy==1.19.5
Pillow==9.0.1
requests==2.27.1
starlette==0.18.0
torch==1.2.0
torchvision==0.4.0
uvicorn==0.17.5
Werkzeug==2.0.3